+++ MOP/src/GameObjects/Vehicles/Managers/Satsuma/SatsumaTrunk.cs
@@ -80,6 +80,7 @@ namespace MOP
 
             if (other.gameObject.GetComponent<ItemHook>() != null && !trunkContent.Contains(other.gameObject))
             {
+                /*
                 if (afterFirstLoad && !storageOpen.Value)
                 {
                     Vector3 newItemPosition = other.gameObject.transform.position;
@@ -88,6 +89,7 @@ namespace MOP
                     other.gameObject.GetComponent<Rigidbody>().velocity = Vector3.zero;
                     return;
                 }
+                */
 
                 trunkContent.Add(other.gameObject);
                 currentMass += other.gameObject.GetComponent<ItemHook>().GetMass();